Added coaddcat.MeasureMultiShears function
git-svn-id: https://desweb.cosmology.illinois.edu/svn/desdm/devel/wl/trunk@3612 2422e80f-9519-0410-a819-8e21e7c86b6d

+++ src/CrudeMeasure.cpp
@@ -312,10 +312,10 @@ void Ellipse::CrudeMeasure(
 {
   std::vector<Pixel> allpix;
   int n = 0;
-  for(int i=0;i<pix.size();i++) n += pix[i].size();
+  for(size_t i=0;i<pix.size();i++) n += pix[i].size();
   allpix.reserve(n);
-  for(int i=0;i<pix.size();i++)
-    for(int j=0;j<pix[i].size();j++)
+  for(size_t i=0;i<pix.size();i++)
+    for(size_t j=0;j<pix[i].size();j++)
       allpix.push_back(pix[i][j]);
   CrudeMeasure(allpix,sigma);
 }
